refactor(core): pipeline-worker normalize-youtube 핸들러 분리
615줄 거대 핸들러 분해 (4/4) — 신규(통합 수집기) / 레거시(영상→댓글 후처리)
두 경로를 별도 파일로 추출. semaphore=3 / since 주입 불필요 동작 보존.

리팩토링 마스터플랜 Phase 2 PR 2-F.
