+++ packages/core/src/queue/pipeline-worker.ts
@@ -1,13 +1,12 @@
 // 파이프라인 Worker 핸들러 -- pipeline 큐 (normalize + persist)
 import type { Job } from 'bullmq';
-import { YoutubeCommentsCollector } from '@ai-signalcraft/collectors';
-import type { YoutubeComment, YoutubeVideo, DataSourceSnapshot } from '@ai-signalcraft/collectors';
-import { updateJobProgress } from '../pipeline';
+import type { DataSourceSnapshot } from '@ai-signalcraft/collectors';
 import { isPipelineCancelled } from '../pipeline/control';
 import { createLogger } from '../utils/logger';
 import { handleClassify } from './pipeline-worker-classify';
 import { handlePersist } from './pipeline-worker-persist';
 import { collectNaverCommentsForArticles } from './pipeline-worker-naver';
+import { splitYoutubeUnifiedResult, collectYoutubeCommentsLegacy } from './pipeline-worker-youtube';
 
 const logger = createLogger('pipeline-worker');
 
@@ -50,154 +49,11 @@ export function createPipelineHandler(): (job: Job) => Promise<any> {
         await collectNaverCommentsForArticles(job, results);
       }
 
-      // normalize-youtube: 일체형 수집기 결과 처리 (영상+댓글 분리)
-      if (job.name === 'normalize-youtube' && results['youtube']) {
-        const videos = (results['youtube'] as { items: YoutubeVideo[] }).items;
-        const allComments: YoutubeComment[] = [];
-
-        for (const video of videos) {
-          allComments.push(...(video.comments ?? []));
-          video.comments = [];
-        }
-
-        results['youtube-videos'] = {
-          source: 'youtube-videos',
-          items: videos,
-          count: videos.length,
-        };
-
-        if (allComments.length > 0) {
-          results['youtube-comments'] = {
-            source: 'youtube-comments',
-            items: allComments,
-            count: allComments.length,
-          };
-        }
-
-        if (dbJobId) {
-          await updateJobProgress(dbJobId, {
-            youtube: {
-              status: 'completed',
-              videos: videos.length,
-              comments: allComments.length,
-            },
-          });
-        }
-      }
-
-      // 하위 호환: 기존 YoutubeVideosCollector로 수집된 결과 처리
-      // NOTE: 이 경로는 "신규 영상 수집 → 댓글 후처리" 전용이며, TTL 재사용된 영상은
-      //       신규 YoutubeCollector(통합형)의 refetchCommentsOnly 경로에서 처리된다.
-      //       따라서 여기서는 since 주입이 불필요 — 모든 영상이 신규 수집 대상이다.
-      if (job.name === 'normalize-youtube' && results['youtube-videos'] && !results['youtube']) {
-        const videos = (
-          results['youtube-videos'] as { items: Array<{ sourceId: string; title?: string }> }
-        ).items;
-        const maxComments = (job.data.maxComments as number) ?? 500;
-        const allComments: YoutubeComment[] = [];
-
-        // 영상별 댓글 수집 진행 추적
-        const videoDetails: Array<{ title: string; status: string; comments: number }> = videos
-          .filter((v) => v.sourceId)
-          .map((v) => ({
-            title: (v.title || v.sourceId).slice(0, 50),
-            status: 'pending',
-            comments: 0,
-          }));
-
-        // 유효한 영상 필터
-        const validVideos: Array<{ index: number; sourceId: string }> = [];
-        for (let i = 0; i < videos.length; i++) {
-          if (videos[i].sourceId) {
-            validVideos.push({ index: i, sourceId: videos[i].sourceId });
-          }
-        }
-
-        // 병렬 수집 -- 동시 3개 영상 댓글 수집 (YouTube API quota 고려)
-        const YT_CONCURRENCY = 3;
-        const updateYtProgress = async () => {
-          if (dbJobId) {
-            await updateJobProgress(dbJobId, {
-              youtube: {
-                status: 'running',
-                videos: videos.length,
-                comments: allComments.length,
-                videoDetails,
-              },
-            });
-          }
-        };
-
-        const collectVideoComments = async (item: { index: number; sourceId: string }) => {
-          const detail = videoDetails[item.index];
-          detail.status = 'running';
-          const collector = new YoutubeCommentsCollector();
-          const videoComments: YoutubeComment[] = [];
-
-          try {
-            for await (const chunk of collector.collect({
-              keyword: item.sourceId,
-              startDate: job.data.startDate ?? '',
-              endDate: job.data.endDate ?? '',
-              maxComments,
-            })) {
-              videoComments.push(...chunk);
-              detail.comments = videoComments.length;
-              await updateYtProgress();
-            }
-            detail.status = 'completed';
-          } catch (err) {
-            // 부분 실패 허용 -- 개별 영상 댓글 실패 시 로깅 후 계속
-            logger.warn(
-              `[youtube-comments] 영상 댓글 수집 실패 (${item.sourceId}):`,
-              err instanceof Error ? err.message : err,
-            );
-            detail.status = 'failed';
-          }
-          return videoComments;
-        };
-
-        // semaphore: YT_CONCURRENCY개씩 배치 처리
-        for (let batchStart = 0; batchStart < validVideos.length; batchStart += YT_CONCURRENCY) {
-          // 배치 시작 전 취소 확인 — 유튜브 댓글 수집 중 즉시 중단
-          if (dbJobId && (await isPipelineCancelled(dbJobId))) {
-            logger.info(
-              `[normalize-youtube] 댓글 수집 중 취소됨 (${allComments.length}건 수집 후)`,
-            );
-            break;
-          }
-
-          const batch = validVideos.slice(batchStart, batchStart + YT_CONCURRENCY);
-          const batchResults = await Promise.allSettled(batch.map(collectVideoComments));
-
-          for (const result of batchResults) {
-            if (result.status === 'fulfilled' && result.value.length > 0) {
-              allComments.push(...result.value);
-            }
-          }
-
-          await job.updateProgress({ commentsCollected: allComments.length });
-          await updateYtProgress();
-        }
-
-        if (allComments.length > 0) {
-          results['youtube-comments'] = {
-            source: 'youtube-comments',
-            items: allComments,
-            count: allComments.length,
-          };
-        }
-
-        // 유튜브 수집 완료 상태로 업데이트
-        if (dbJobId) {
-          await updateJobProgress(dbJobId, {
-            youtube: {
-              status: 'completed',
-              videos: videos.length,
-              comments: allComments.length,
-              videoDetails,
-            },
-          });
+      // normalize-youtube: 일체형(신규) 또는 레거시 YoutubeVideosCollector 결과 처리
+      if (job.name === 'normalize-youtube') {
+        const unifiedHandled = await splitYoutubeUnifiedResult(job, results);
+        if (!unifiedHandled) {
+          await collectYoutubeCommentsLegacy(job, results);
         }
       }
